Rename package from "cluster" to "funk"
"cluster" is too generic and clients end up calling the variable
"cluster" for obvious reasons. Now there's less stuttering
(Cluster cluster.Cluster popped up in the demo server and it's not
pretty) The end result is declarations like "cluster funk.Cluster"
and that sounds a lot better when reading it out loud.