Liveness checker
Implement a separate liveness checker that can be used alongside Raft.
The Raft heartbeats uses a lot of bandwidth so a much simpler liveness
test reduces the bandwidth requirements for clusters.

The timeouts for Raft could possibly be tuned up as well. Leader election
would still be quite quick but changes in cluster size without an election
is faster.