Use own liveness tester rather than custom Raft
Drop the custom Raft implementation and use the more aggressive liveness
test. This uses considerably less bandwidth but duplicates some
of the functionality already in the Raft library. Detection is quicker (since it's
more aggressive) and it might need a tune before
it is ready for prime time.

Added new methods in RaftNode (EnableNode/DisableNode), reintroduced
Clear method in liveness checker.